@@ -19,8 +19,8 @@
         }
     ],
     "require": {
-        "laravel/framework": "~5.1|~6.0|~7.0|~8.0|~9.0|^10.0|^11.0",
-        "php": "^7.1|^8.0|^8.1",
+        "laravel/framework": "^10.0|^11.0",
+        "php": "^8.0|^8.1",
         "ext-gd": "*"
     },
     "require-dev": {
